spectrum analyzer improvements

+++ synth/src/dsp/math.hpp
@@ -19,8 +19,8 @@
 #define __DSP_MATH_HPP_
 
 namespace dsp {
-    inline float fastAtan(float inp){
-        return inp / (1.0f + 0.28f * inp * inp);
+    inline double fastAtan(double inp){
+        return inp / (1.0 + 0.28 * inp * inp);
     }
 }
